Add utilities for extracting form validation errors from mutation response.
If caught error is of StructError type, we look for any keys that match known errors schema.
If some matching errors are found, we can use their keys to display them right next to invalid form field.
If no matches are found, we can show a toast with error message instead, so that there is some feedback for the user