Add `Fetch` type to api helpers (fetch 1 record by id)
Currently all book mutations return `Book` type (w/o volumes data). So it can be handy to fetch a `BookItem` (with full data) by id.

Note: all book items are currently always present in the main query cache, but that's not guaranteed to be the case. Cache may not contain book item with given id, because of pagination, or because it was just created etc